Working role authentication with data base, working book api

+++ src/main/java/com/myBookSchelf/BookSchelf/model/User.java
@@ -19,7 +19,7 @@ public class User {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
 //    @Column(name = "user_id")
-    private int id;
+    private long id;
 
     private String name;
 @Column(nullable = false, unique = true)
@@ -34,8 +34,6 @@ private String username;
 @ManyToMany(fetch = FetchType.EAGER, cascade = CascadeType.ALL)
 @JoinTable(name = "users_roles", joinColumns= @JoinColumn(name = "user_id", referencedColumnName ="id"),
 inverseJoinColumns = @JoinColumn (name = "role_id", referencedColumnName = "id"))
-
-
 private Set<Role> roles;
